feat: Extract `AbstractCmi5` superclass.
